+++ tests/testcase.py
@@ -3,10 +3,8 @@ import os
 import platform
 import shutil
 import sys
-import time
 import unittest
 import uuid
-from functools import wraps
 
 import psutil
 from psutil import AccessDenied, NoSuchProcess
@@ -16,9 +14,7 @@ from pyngrok import process
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2021, Alex Laird"
-__version__ = "5.0.4"
-
-from pyngrok.exception import PyngrokNgrokError
+__version__ = "5.0.5"
 
 logger = logging.getLogger(__name__)
 ngrok_logger = logging.getLogger("{}.ngrok".format(__name__))
@@ -34,6 +30,7 @@ class NgrokTestCase(unittest.TestCase):
         conf.DEFAULT_NGROK_CONFIG_PATH = config_path
         self.pyngrok_config = conf.get_default()
         self.pyngrok_config.config_path = conf.DEFAULT_NGROK_CONFIG_PATH
+        self.pyngrok_config.reconnect_session_retries = 10
 
         installer.DEFAULT_RETRY_COUNT = 1
 
@@ -71,39 +68,3 @@ class NgrokTestCase(unittest.TestCase):
         return "pyngrok-{}-{}-{}-{}{}-tcp".format(uuid.uuid4(), platform.system(),
                                                   platform.python_implementation(), sys.version_info[0],
                                                   sys.version_info[1]).lower()
-
-
-def retry_connection_reset():
-    """
-    This decorator can be applied to tests that are flaky due to ngrok failing to reconnect the remote session. The EOF
-    error is seen when the ngrok remote host says "failed to reconnect session", which may mean ngrok itself is having
-    issues (unrelated to the test, retry), or it can be seen when ngrok is throttling connections from the test account
-    (too many consecutive or concurrent connections, slow down and retry). In either case, this decorator will apply a
-    delay, backoff, and then retry the test.
-    """
-
-    def decorator(func):
-
-        @wraps(func)
-        def inner(*args, **kwargs):
-            delay = 2
-            retries = 3
-            while retries > 1:
-                try:
-                    return func(*args, **kwargs)
-                except PyngrokNgrokError as e:
-                    # Raise the exception if it's not the specific edge case we're looking for
-                    if e.ngrok_error != "EOF":
-                        raise e
-
-                    logger.warning("ngrok reset our connection, retrying in {} seconds ...".format(delay))
-                    time.sleep(delay)
-
-                    retries -= 1
-                    delay *= 2
-
-            return func(*args, **kwargs)
-
-        return inner
-
-    return decorator